+++ Dominio/TipoHabitacion.cs
@@ -12,7 +12,7 @@ namespace Dominio
 
         public string Nombre { get; set; }
 
-        public float Precio { get; set; }
+        public Decimal Precio { get; set; }
 
         public override string ToString()
         {
